// src/app/models/camiseta.ts
export interface Camiseta {
    id: number;
    equipo: string;
    anho: string; 
    tipo: string;
    zona: string;
    pais: string;
    clave: string;
    equipacion: string;
    marca: string;
    talla: string;
    nombre: string;
    numero: string;
    imagen1: string;
    imagen2: string;
}